+++ yaml-config-updater/src/main/java/ru/vyarus/yaml/updater/update/EnvSupport.java
@@ -3,6 +3,7 @@ package ru.vyarus.yaml.updater.update;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.HashMap;
 import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -19,69 +20,71 @@ import java.util.stream.Collectors;
  * in format {@code ${va}} are often used directly in the configuration (to be replaced at runtime with environment
  * variables).
  * <p>
- * Variable names are case sensitive. Unknown variables are not replaced. Null values replaced with empty.
+ * Variable names are case-sensitive. Unknown variables are not replaced. Null values replaced with empty.
  *
  * @author Vyacheslav Rusakov
  * @since 08.06.2021
  */
-public final class EnvSupport {
-    private static final Logger LOGGER = LoggerFactory.getLogger(EnvSupport.class);
+public class EnvSupport {
+    private final Logger logger = LoggerFactory.getLogger(EnvSupport.class);
 
-    private EnvSupport() {
-    }
+    private final String prefix;
+    private final String postfix;
+    private final Map<String, String> env;
 
-    /**
-     * Replace variables in the provided text.
-     *
-     * @param text text to process
-     * @param env  variables map
-     * @return text with replaced known variables
-     */
-    public static String apply(final String text,
-                               final Map<String, String> env) {
+    private final Map<String, String> applied = new HashMap<>();
+
+    public EnvSupport(final Map<String, String> env) {
         // non standard vars identity used because:
         // - real config could rely on environment variables
         // - not processed file would remain valid yaml because placeholder would be interpreted as comment
-        return apply(text, "#{", "}", env);
+        this("#{", "}", env);
     }
 
-    /**
-     * Replace variables in the provided text.
-     *
-     * @param text    text to process
-     * @param prefix  variable prefix
-     * @param postfix variable postfix
-     * @param env     variables map
-     * @return text with replaced known variables
-     */
-    public static String apply(final String text,
-                               final String prefix,
-                               final String postfix,
-                               final Map<String, String> env) {
-        if (text == null || text.isEmpty() || env == null || env.isEmpty()) {
-            return text;
-        }
+    public EnvSupport(final String prefix,
+                       final String postfix,
+                       final Map<String, String> env) {
         if (prefix == null) {
             throw new IllegalArgumentException("Variable prefix required");
         }
         if (postfix == null) {
             throw new IllegalArgumentException("Variable postfix required");
         }
+        this.prefix = prefix;
+        this.postfix = postfix;
+        this.env = env;
+    }
 
-        if (LOGGER.isDebugEnabled()) {
-            LOGGER.debug("Replacing variables in format '{}name{}' from: \n{}", prefix, postfix,
+    /**
+     * Replace variables in the provided text.
+     *
+     * @param text text to process
+     * @return text with replaced known variables
+     */
+    public String apply(final String text) {
+        if (text == null || text.isEmpty() || env == null || env.isEmpty()) {
+            return text;
+        }
+
+        if (logger.isDebugEnabled()) {
+            logger.debug("Replacing variables in format '{}name{}' from: \n{}", prefix, postfix,
                     env.entrySet().stream()
                             .map(entry -> "    " + entry.getKey() + "=" + entry.getValue())
                             .collect(Collectors.joining("\n")));
         }
-        return replace(text, prefix, postfix, env);
+        return replace(text);
+    }
+
+    /**
+     * @return applied variables in the latest execution or empty map
+     */
+    public Map<String, String> getApplied() {
+        return new HashMap<>(applied);
     }
 
     @SuppressWarnings("checkstyle:IllegalIdentifierName")
-    private static String replace(final String text,
-                                  final String prefix,
-                                  final String postfix,
-                                  final Map<String, String> env) {
+    private String replace(final String text) {
+        applied.clear();
         String res = text;
         for (Map.Entry<String, String> entry : env.entrySet()) {
             final String var = prefix + entry.getKey() + postfix;
@@ -97,8 +100,9 @@ public final class EnvSupport {
                 cntr++;
             }
             if (cntr > 0) {
-                LOGGER.debug("    {} ({}) replaced with: {}", var, cntr, replacement);
+                logger.debug("    {} ({}) replaced with: {}", var, cntr, replacement);
                 res = matcher.replaceAll(replacement);
+                applied.put(entry.getKey(), value);
             }
         }
         return res;