+++ yaml-config-updater/src/test/groovy/ru/vyarus/yaml/updater/BackupTest.groovy
@@ -2,6 +2,7 @@ package ru.vyarus.yaml.updater
 
 import ru.vyarus.yaml.updater.listen.UpdateListener
 import ru.vyarus.yaml.updater.listen.UpdateListenerAdapter
+import ru.vyarus.yaml.updater.report.ReportPrinter
 import spock.lang.TempDir
 
 import java.nio.file.Files
@@ -26,11 +27,26 @@ class BackupTest extends AbstractTest {
 
         when: "updating"
         UpdLst list = new UpdLst()
-        YamlUpdater.create(current, update).backup(true).listen(list).update()
+        def report = YamlUpdater.create(current, update).backup(true).listen(list).update()
 
         then: "backup created"
         dir.list().size() == 3
         list.backup != null
+
+        and: "report correct"
+        ReportPrinter.print(report) == """Configuration: /tmp/CONFIG.yml (198 bytes, 23 lines)
+Updated from source of 385 bytes, 40 lines
+Resulted in 331 bytes, 36 lines
+
+\tAdded from new file:
+\t\tprop1/prop1.3                            9  | prop1.3: 1.3
+\t\tprop11                                   12 | prop11:
+\t\tprop2/obj[0]/three                       31 | three: 3
+\t\tprop3                                    39 | prop3:
+
+\tBackup created: BACKUP
+""".replace("/tmp/CONFIG.yml", current.getAbsolutePath())
+        .replace("BACKUP", list.backup.name)
     }
 
     static class UpdLst extends UpdateListenerAdapter {